Makes Modbus close operation asynchronous
Updates the Modbus `close` operation to be asynchronous, ensuring proper disconnection from the Modbus context.

This change prevents blocking the main thread during disconnection and ensures that the disconnection process is handled correctly within the asynchronous runtime.

Also, the websocket message struct is introduced to distinguish between new connections and messages to be sent.
Broadcast function also optimized, the message will be directly sent if only one client is connected.