Refactors database schema and adds logging
Updates the database schema for settings to use a key-value pair instead of name, sequence and description, and adds a unique index.

Implements time-partitioned logging with a trigger function to create monthly partitions.
The logs table is now partitioned by time for better performance and manageability.

The t_logs.user_id field is now optional to handle cases where a user is not associated with a log entry.

Adds timestamping to user and settings creation.
Creates a default "sys" user if one does not exist.